Reduced footprint and some code cleanup

+++ NexaLightSwitchPacket.cpp
@@ -32,13 +32,20 @@ NexaLightSwitchPacket::~NexaLightSwitchPacket()
     delete ook_msg;
 }
 
-void NexaLightSwitchPacket::setBit(WireMessage *ook_msg, unsigned int ms_high, unsigned int ms_low)
+void NexaLightSwitchPacket::setBit(
+        WireMessage *ook_msg,
+        uint16_t ms_high,
+        uint16_t ms_low)
 {
     ook_msg->ms_delay_high = ms_high;
     ook_msg->ms_delay_low = ms_low;
 }
 
-void NexaLightSwitchPacket::appendData(WireMessage *ook_msg, unsigned int *msg_index, unsigned int data, unsigned int data_width)
+void NexaLightSwitchPacket::appendData(
+        WireMessage *ook_msg,
+        uint16_t *msg_index,
+        uint16_t data,
+        uint8_t data_width)
 {
     for (int i = 0; i < data_width; i++) {
         setBit(&ook_msg[(*msg_index)++], T_SHORT, T_LONG);
@@ -71,12 +78,12 @@ const WireMessage *NexaLightSwitchPacket::getWireData() const
     return ook_msg;
 }
 
-unsigned int NexaLightSwitchPacket::getWireDataSize() const
+uint16_t NexaLightSwitchPacket::getWireDataSize() const
 {
     return NEXA_LIGHT_SWITCH_MESSAGE_SIZE;
 }
 
-unsigned int NexaLightSwitchPacket::getTransmitionRepeats() const
+uint8_t NexaLightSwitchPacket::getTransmitionRepeats() const
 {
     return 3;
 }
@@ -84,12 +91,12 @@ unsigned int NexaLightSwitchPacket::getTransmitionRepeats() const
 //////////////////////////////////////////////////////////////////////////
 // Setters
 
-void NexaLightSwitchPacket::setHouse(unsigned int house)
+void NexaLightSwitchPacket::setHouse(uint8_t house)
 {
     this->house = house;
 }
 
-void NexaLightSwitchPacket::setUnit(unsigned int unit)
+void NexaLightSwitchPacket::setUnit(uint8_t unit)
 {
     this->unit = unit;
 }
